Добавление разграничения прав

+++ src/RisovavitiApi/Controllers/VersionProjectController.cs
@@ -1,8 +1,11 @@
-﻿using DomainModel.Integration.CanvasOperation;
+﻿using DomainModel.Exceptions;
+using DomainModel.Integration.CanvasOperation;
+using DomainModel.ResultsRequest;
 using DomainModel.ResultsRequest.Canvas;
 using Logic.Interface;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using RisovavitiApi.Model.Interfaces;
 using RisovavitiApi.UserOperate;
 
 namespace RisovavitiApi.Controllers
@@ -15,14 +18,17 @@ namespace RisovavitiApi.Controllers
 		IFabricOperateVersionProject _operate;
 		IBuilderGetterVerionsByParent _builder;
 		IEditVersionProject _editVersionProject;
+		IDefinitionerOfPermissionByHttpContext _definitionerOfPermission;
 
 		public VersionProjectController(IFabricOperateVersionProject operate, 
 										IBuilderGetterVerionsByParent getterVerionsByParent,
-										IEditVersionProject editVersion) 
+										IEditVersionProject editVersion,
+										IDefinitionerOfPermissionByHttpContext definitionerOfPermission) 
 		{
 			_operate = operate;
 			_builder = getterVerionsByParent;
 			_editVersionProject = editVersion;
+			_definitionerOfPermission = definitionerOfPermission;
 		}
 
 		[HttpGet("get/{id}")]
@@ -31,6 +37,10 @@ namespace RisovavitiApi.Controllers
 			try
 			{
 				return await TryGetById(id);
+			}
+			catch (NotPermissionExeption)
+			{
+				return Unauthorized();
 			}
 			catch (Exception ex) 
 			{
@@ -42,6 +52,11 @@ namespace RisovavitiApi.Controllers
 		{
 			var getter = _operate.CreateGetter(UserGetterByContext.GetUserIntegration(HttpContext));
 			var vertion = await getter.GetVersionProjectByIdAsync(id);
+			PermissionResult permissionResult = await _definitionerOfPermission.GetPermissionResult(vertion);
+			if(!permissionResult.Read ?? false)
+			{
+				throw new NotPermissionExeption();
+			}
 			return Ok(vertion);
 		}
 
@@ -76,6 +91,13 @@ namespace RisovavitiApi.Controllers
 		[HttpPost("edit")]
 		public async Task<IActionResult> Edit([FromBody] VerstionProjectEditResutl newVerstion)
 		{
+			var getter = _operate.CreateGetter(UserGetterByContext.GetUserIntegration(HttpContext));
+			var vertion = await getter.GetVersionProjectByIdAsync(newVerstion.VerstionId);
+			PermissionResult permission = await _definitionerOfPermission.GetPermissionResult(vertion);
+			if(!permission.Edit ?? false)
+			{
+				return Unauthorized();
+			}
 			await _editVersionProject.Edit(newVerstion);
 			return Ok();
 		}
@@ -83,6 +105,13 @@ namespace RisovavitiApi.Controllers
 		[HttpPost("delet/{id}")]
 		public async Task<IActionResult> Delet(int id)
 		{
+			var getter = _operate.CreateGetter(UserGetterByContext.GetUserIntegration(HttpContext));
+			var vertion = await getter.GetVersionProjectByIdAsync(id);
+			PermissionResult permission = await _definitionerOfPermission.GetPermissionResult(vertion);
+			if (!permission.Edit ?? false)
+			{
+				return Unauthorized();
+			}
 			var adder = _operate.CreateAdder(UserGetterByContext.GetUserIntegration(HttpContext));
 			await adder.DeleteVertionProjectAsync(id);
 			return Ok();
